sequelize使用关联关系，登陆成功组件自动刷新，登陆增加有效时间判断

+++ server/src/resolvers.js
@@ -1,11 +1,12 @@
 const { jwt } = require('./utils');
 const { Authorize } = require('./decorator.js');
 const { PubSub } = require('apollo-server');
+const isEmail = require('isemail');
 
 const pubsub = new PubSub();
 const MEMBER_LEAVE = 'MEMBER_LEAVE';
 
-module.exports = {
+var resolvers = {
     Query: {
         me: async (_, __, { dataSources }) => dataSources.userAPI.findOrCreateUser()
     },
@@ -16,83 +17,110 @@ module.exports = {
     },
     Mutation: {
         login: async (_, { email }, { dataSources }) => {
-            const user = await dataSources.userAPI.findOrCreateUser({ email });
+            if (!email || !isEmail.validate(email)) {
+                return {
+                    code: -2,
+                    success: false,
+                    message: '邮箱格式不正确',
+                    token: ''
+                };
+            }
+            const user = await dataSources.userAPI.findOrCreateUser(email);
             if (user)
                 return {
+                    code: 0,
                     success: true,
                     message: '登陆成功',
                     token: jwt.sign({ email })
                 };
             return {
+                code: -3,
                 success: false,
-                message: '邮箱格式不正确',
+                message: '未知错误',
                 token: ''
             };
         },
-        updateMyName: Authorize(async (_, { name }, { dataSources }) => {
-            await dataSources.userAPI.updateNameByUser(name);
+        @Authorize
+        updateMyName: async (_, { name }, { dataSources }) => {
+            await dataSources.userAPI.updateMyName(name);
             return {
+                code: 0,
                 success: true,
                 message: '更新成功'
             };
-        }),
-        addWorkSpace: Authorize(async (_, { name }, { dataSources }) => {
+        },
+        @Authorize
+        addWorkSpace: async (_, { name }, { dataSources }) => {
             const workSpace = await dataSources.workSpaceAPI.findOrCreateWorkSpace({ name });
             return {
+                code: 0,
                 success: true,
                 message: '更新成功',
-                workSpace: workSpace.dataValues
+                workSpace: workSpace
             };
-        }),
-        updateWorkSpaceName: Authorize(async (_, { WorkSpace }, { dataSources }) => {
+        },
+        @Authorize
+        updateWorkSpaceName: async (_, { WorkSpace }, { dataSources }) => {
             var res = await dataSources.workSpaceAPI.updateWorkSpaceName(WorkSpace);
             if (res[0]) {
                 return {
+                    code: 0,
                     success: true,
                     message: '更新成功',
                     workSpace: { id: WorkSpace.id }
                 };
             } else {
                 return {
+                    code: -2,
                     success: false,
                     message: '更新失败',
                     workSpace: null
                 };
             }
-        }),
-        deleteWorkSpace: Authorize(async (_, { id }, { dataSources }) => {
+        },
+        @Authorize
+        deleteWorkSpace: async (_, { id }, { dataSources }) => {
             await dataSources.workSpaceAPI.deleteWorkSpace(id);
             return {
+                code: 0,
                 success: true,
                 message: '删除成功'
             };
-        }),
-        addWorkSpaceMembers: Authorize(async (_, { id, emails }, { dataSources }) => {
-            await dataSources.workSpaceAPI.addWorkSpaceMembers(id, emails);
+        },
+        @Authorize
+        addWorkSpaceMembers: async (_, { id, emails }, { dataSources }) => {
+			emails = emails.filter(email => isEmail.validate(email));
+            if (emails.length !== 0) {
+                await dataSources.workSpaceAPI.addWorkSpaceMembers(id, emails);
+            }
             return {
+                code: 0,
                 success: true,
                 message: '更新成功',
                 workSpace: { id }
             };
-        }),
-        deleteWorkSpaceMembers: Authorize(async (_, { id, emails }, { dataSources }) => {
-            pubsub.publish(MEMBER_LEAVE, { memberLeave: { id, emails } });
-            var users = await dataSources.workSpaceAPI.deleteWorkSpaceMembers(id, emails);
+        },
+        @Authorize
+        deleteWorkSpaceMembers: async (_, { id, emails }, { dataSources }) => {
+			//pubsub.publish(MEMBER_LEAVE, { memberLeave: { id, emails } });
+			emails = emails.filter(email => isEmail.validate(email));
+			if (emails.length !== 0) {
+				await dataSources.workSpaceAPI.deleteWorkSpaceMembers(id, emails);
+			}
             return {
+                code: 0,
                 success: true,
                 message: '更新成功',
                 workSpace: { id }
             };
-        })
+        }
     },
     WorkSpace: {
         owner: async (workSpace, __, { dataSources }) => {
             return await dataSources.userAPI.getUserById(workSpace.userId);
         },
         members: async (workSpace, __, { dataSources }) => {
-            const users = await dataSources.workSpaceAPI.getMemberIdsById(workSpace.dataValues.id);
-            const members = await dataSources.userAPI.getUserByIds(users.map(item => item.dataValues.userId));
-            return members.map(item => item.dataValues);
+            return await dataSources.workSpaceAPI.getMembersById(workSpace.id);
         }
     },
     User: {
@@ -105,7 +133,7 @@ module.exports = {
     },
     UserResponse: {
         me: async (_, __, { dataSources }) => {
-            return dataSources.userAPI.getMy();
+            return await dataSources.userAPI.getMe();
         }
     },
     WorkSpaceResponse: {
@@ -114,3 +142,5 @@ module.exports = {
         }
     }
 };
+
+module.exports = resolvers;
